Pull reused functions out of timeConductor e2e tests into appActions
